requirements.txt
aiohttp           3.7.4.post0
asgiref           3.3.4
async-timeout     3.0.1
attrs             20.3.0
bcrypt            3.2.0
certifi           2020.12.5
cffi              1.14.5
chardet           4.0.0
click             7.1.2
cryptography      3.4.7
discord.py        1.7.1
Django            3.2
Flask             1.1.2
future            0.18.2
idna              2.10
itsdangerous      1.1.0
Jinja2            2.11.3
MarkupSafe        1.1.1
multidict         5.1.0
netmiko           3.4.0
ntc-templates     2.0.0
paramiko          2.7.2
pip               20.2.3
pycparser         2.20
PyJWT             2.0.1
PyNaCl            1.4.0
pyserial          3.5
python-webex-bot  0.854
pytz              2021.1
requests          2.25.1
requests-toolbelt 0.9.1
scp               0.13.3
setuptools        49.2.1
six               1.15.0
sqlparse          0.4.1
tenacity          7.0.0
textfsm           1.1.0
typing-extensions 3.7.4.3
urllib3           1.26.4
webexteamsbot     0.1.4.2
webexteamssdk     1.6
Werkzeug          1.0.1
yarl              1.6.3